Refactored UIViewWrapper
